Use safe execution for registry query on Windows
due to an issue in Deno about quoting some special characters
